creation Compétences page and skillCard compenant
