+++ data/Skills.ts
@@ -1,14 +1,4 @@
-export type Skill = {
-  id: string;
-  name: string;
-  icon: string;
-  description: string;
-};
-
-export type SkillCategory = {
-  category: string;
-  skills: Skill[];
-};
+import { SkillCategory } from "@/types/SkillTypes";
 
 export const skills: SkillCategory[] = [
   {
