+++ src/app/(site)/competences/page.tsx
@@ -1,16 +1,19 @@
+"use client";
+import { skills } from "../../../../data/Skills";
+import SkillCard from "../components/SkillCard";
+
 export default function Competences() {
   return (
     <div>
-      <h1>Competences</h1>
-      <p>Here you can find the competences of the project.</p>
-      <ul>
-        <li>Competence 1</li>
-        <li>Competence 2</li>
-        <li>Competence 3</li>
-         <li>Competence 1</li>
-        <li>Competence 2</li>
-        <li>Competence 3</li>
-      </ul>
+      <h1 className="font-reddit blue-text text-2xl md:text-6xl font-semibold text-center mt-5 md:mt-30 mb-5 md:mb-20">
+        Compétences
+      </h1>
+      {skills.map((category, index) => (
+        <div key={index}>
+          <h2 className="font-reddit blue-text text-2xl md:text-4xl font-semibold text-center mt-5 md:mt-30 mb-5 md:mb-10">{category.category}</h2>
+          <SkillCard skills={category.skills} />
+        </div>
+      ))}
     </div>
   );
 }
